+++ src/main/scala/tech/sourced/api/udf/ExtractUASTsUDF.scala
@@ -35,7 +35,6 @@ object ExtractUASTsUDF extends CustomUDF {
 
   def extractUsingBblfsh(bblfshClient: BblfshClient, path: String, content: Array[Byte], lang: String): Array[Byte] = {
     //FIXME(bzz): not everything is UTF-8 encoded :/
-    println(s"Parsing path:'$path' with content:'$content' using lang:'$lang'")
     val parsed = bblfshClient.parse(path, content = new String(content, StandardCharsets.UTF_8), lang = lang)
     if (parsed.errors.isEmpty) {
       parsed.uast.get.toByteArray
